combine sync/protocol files
